+++ dev/test/so_5/bench/ping_pong/main.cpp
@@ -283,12 +283,13 @@ show_result(
 
 		const unsigned int total_msg_count = cfg.m_request_count * 2;
 
-		double price = static_cast< double >( total_msec ) / total_msg_count / 1000.0;
+		double price = static_cast< double >( total_msec ) /
+				double(total_msg_count) / 1000.0;
 		double throughtput = 1 / price;
 
 		benchmarks_details::precision_settings_t precision{ std::cout, 10 };
 		std::cout <<
-			"total time: " << total_msec / 1000.0 << 
+			"total time: " << double(total_msec) / 1000.0 << 
 			", messages sent: " << total_msg_count <<
 			", price: " << price <<
 			", throughtput: " << throughtput << std::endl;
